sing in and login correctly, sectionPanel.innerHTML add, some problems with that
