--- js/Class/Client.js
@@ -1,12 +1,0 @@
-
-export const testClient = console.log("Client File Loaded");
-
-export default class Client {
-  constructor(fullName, email, country, phoneNumber, processStatus) {
-    this.fullName = fullName;
-    this.email = email;
-    this.phoneNumber = phoneNumber;
-    this.country = country;
-    this.processStatus = processStatus;
-  }
-}
