--- js/Class/Promotor.js
@@ -1,10 +1,0 @@
-
-export const testPromotor = console.log("Promotor File Loaded");
-
-export default class Promotor {
-    constructor(userName, password) {
-      this.userName = userName;
-      this.password = password;
-    }
-  }
-
